requirements.txt
numpy==2.2.4
PyQt5==5.15.11
matplotlib==3.10.1
Pillow
pandas==2.2.3
tensorflow